requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
gunicorn
pytest
pytest-cov
flake8


